MediSolution
The objectives of creating “Medisolution” are to quickly request an appointment with the
trusted doctor according to location and category, share scans/reports - all from the comfort
of your own home with MediSolution. There’s no need to travel large distances for a quick
consultation. Using the website gives flexibility to save money and time.It can search for
medicine according to the user’s preference. Users can also review doctors. As doctors are
having hard times, they are not sitting in their chambers regularly. Most of the doctors are
not visiting their chambers at all due to their Covid-19 duty. “Medisolution” is a platform
aiming to reduce that miscommunication.